[REFACTOR]: Separate slider hook from component
